pub sub for pos and vel working, need to control pos publisher

+++ Arduino code/v4_4_1_six_float_msg_light_test/MotorStepper.h
@@ -11,8 +11,8 @@
 class MotorStepper {
 	public:
 		MotorStepper(int pulse, int direct, int lower, int upper,float stepSize, float multipler);
-		int getPos();
-		int getVel();
+		float getPos();
+		float getVel();
 		void setVel(float incoming_vel);
 		void pulse();
 		void checkStep(unsigned long current_time);
